angler: suppress unused parameter warning
It looks like these parameters are used by the proprietary
mm-camera-daemon blob, so we can't really get rid of them easily without
rewriting the whole code and putting hands on the binary (which would
likely bring a lot of issues).
Let's just suppress the warnings and we should be just fine.

Change-Id: I8891749652fcbf3103bc1562f57fdcd9dbe6cdc5